+++ deploy/cloud/operator/internal/dynamo/component_common.go
@@ -28,3 +28,18 @@ func ComponentDefaultsFactory(componentType string) ComponentDefaults {
 		return &BaseComponentDefaults{}
 	}
 }
+
+// BaseComponentDefaults provides common defaults shared by all components
+type BaseComponentDefaults struct{}
+
+func (b *BaseComponentDefaults) GetBaseContainer(backendFramework BackendFramework) (corev1.Container, error) {
+	return b.getCommonContainer(), nil
+}
+
+func (b *BaseComponentDefaults) getCommonContainer() corev1.Container {
+	container := corev1.Container{
+		Name: "main",
+	}
+
+	return container
+}
