+++ deploy/cloud/operator/internal/dynamo/component_frontend.go
