+++ deploy/cloud/operator/internal/dynamo/component_worker.go
